Finish keep-alive packet

+++ elytra-sdk/src/main/kotlin/io/elytra/sdk/network/protocol/message/play/KeepAliveMessage.kt
@@ -2,4 +2,4 @@ package io.elytra.sdk.network.protocol.message.play
 
 import com.flowpowered.network.Message
 
-data class KeepAliveMessage(val id: Int) : Message
+data class KeepAliveMessage(val id: Long) : Message
